MISUV-9576:   * code tidy up => FinancialDetailsHipConnectorSpec / other refactoring applied;
